+++ src/Constants.ts
@@ -1,5 +1,7 @@
 export const DEFAULT_OFFSET = 50
 export const DDP_LOG_RETENTION_LIMIT = 5000
+export const GITHUB_REPOSITORY_URL =
+  'https://github.com/leonardoventurini/meteor-devtools-evolved'
 
 export const DEVELOPMENT = import.meta.env.MODE === 'development'